Establish admin-only api calls

+++ models/user.js
@@ -6,12 +6,22 @@ const cryptoUtil = require('../helpers/cryptoUtil.js');
 const collectionName = 'users';
 
 exports.all = (callback) => {
+  if (!req.decoded.isAdmin) {
+    const error = new Error('Access Denied: Admins only');
+    error.status = 401;
+    return next(error);
+  }
   mongoUtil.getDb().collection(collectionName).find().toArray((err, result) => {
     callback(err, result);
   });
 };
 
 exports.get = (id, callback) => {
+  if (!req.decoded.isAdmin) {
+    const error = new Error('Access Denied: Admins only');
+    error.status = 401;
+    return next(error);
+  }
   mongoUtil.getDb().collection(collectionName).findOne({ _id: ObjectId(id) }, (err, result) => {
     callback(err, result);
   });
@@ -30,6 +40,11 @@ exports.new = (data, callback) => {
 };
 
 exports.update = (id, data, callback) => {
+  if (!req.decoded.isAdmin) {
+    const error = new Error('Access Denied: Admins only');
+    error.status = 401;
+    return next(error);
+  }
   const hashResult = cryptoUtil.saltHashPassword(data.password);
   mongoUtil.getDb().collection(collectionName).updateOne({ _id: ObjectId(id) }, {
     username: data.username,
@@ -41,6 +56,11 @@ exports.update = (id, data, callback) => {
 };
 
 exports.delete = (id, callback) => {
+  if (!req.decoded.isAdmin) {
+    const error = new Error('Access Denied: Admins only');
+    error.status = 401;
+    return next(error);
+  }
   mongoUtil.getDb().collection(collectionName).deleteOne({ _id: ObjectId(id) }, (err) => {
     callback(err);
   });
